+++ attributes/default.rb
@@ -70,6 +70,12 @@ default['openstack']['image']['registry']['workers'] = [8, node['cpu']['total'].
 # Return the URL that references where the data is stored on the backend.
 default['openstack']['image']['api']['show_image_direct_url'] = 'False'
 
+# Supported values for the 'container_format' image attribute
+default['openstack']['image']['api']['container_formats'] = %w(ami ari aki bare ovf ova)
+
+# Supported values for the 'disk_format' image attribute
+default['openstack']['image']['api']['disk_formats'] = %w(ami ari aki vhd vmdk raw qcow2 vdi iso)
+
 default['openstack']['image']['api']['auth']['version'] = node['openstack']['api']['auth']['version']
 default['openstack']['image']['registry']['auth']['version'] = node['openstack']['api']['auth']['version']