+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This file is used to list changes made in each version of cookbook-openstack-ima
 * Add oslo.messaging attributes for api and registry conf files
 * Add support for upload of vhd vmdk vdi iso raw disk formats
 * Allow some attributes for cinder storage backend to be configurable
+* Add support for multiple rabbit mq hosts
 
 ## 9.2.0
 * python_packages database client attributes have been migrated to the -common cookbook